requirements.txt
cmake==3.24.1.1
face-recognition==1.3.0
scikit-learn==1.1.2
opencv-python==4.6.0.66
dlib==19.24.0
face-recognition-models==0.3.0
click==8.1.3
colorama==0.4.5
joblib==1.2.0
numpy==1.23.4
Pillow==9.2.0
scipy==1.9.2
threadpoolctl==3.1.0